Module3 Task 1,2

+++ src/Module3/Task2/Arithmetic.java
@@ -4,7 +4,7 @@ public class Arithmetic {
 
     public int add(Integer a, Integer b){
 
-        return (int)a + b;
+        return a + b;
     }
 
 }
